feat: added contact component
